+++ project-management/src/main/java/life/qbic/projectmanagement/application/sample/qualitycontrol/QualityControlStorageException.java
@@ -1,7 +1,7 @@
-package life.qbic.projectmanagement.application.api;
+package life.qbic.projectmanagement.application.sample.qualitycontrol;
 
 import java.io.Serial;
-import life.qbic.projectmanagement.domain.model.project.qualityControl.QualityControl;
+import life.qbic.projectmanagement.domain.model.sample.qualitycontrol.QualityControl;
 
 /**
  * <b>QualityControlStorageException</b>
